+++ index.js
@@ -9,7 +9,8 @@ module.exports = {
     'no-mutating-assign': require('./rules/no-mutating-assign'),
     'no-mutation': require('./rules/no-mutation'),
     'no-this': require('./rules/no-this'),
-    'no-throw': require('./rules/no-throw')
+    'no-throw': require('./rules/no-throw'),
+    'no-unused-expression': require('./rules/no-unused-expression')
   },
   configs: {
     recommended: {
@@ -22,7 +23,8 @@ module.exports = {
         'fp/no-mutating-assign': 'error',
         'fp/no-mutation': 'error',
         'fp/no-this': 'error',
-        'fp/no-throw': 'error'
+        'fp/no-throw': 'error',
+        'fp/no-unused-expression': 'error'
       }
     }
   }